+++ src/Scene.h
@@ -21,7 +21,7 @@ public:
 	virtual void render(sf::RenderWindow& window) = 0;
 	virtual bool clearState() = 0;
 
-	sf::Vector2i MouseInput;
+	sf::Vector2f MouseInput;
 
 		//stores a unique pointer to the objects within the vector
 	std::vector<std::shared_ptr<ObjectClass>> state_obj;
